+++ GeneratorDuty/CallBackKeyboards/Duty/DutyAccept.cs
@@ -1,24 +1,23 @@
-﻿using GeneratorDuty.Commands;
-using GeneratorDuty.Common;
+﻿using GeneratorDuty.Common;
 using GeneratorDuty.CustomRights;
 using GeneratorDuty.Extensions;
 using GeneratorDuty.Models;
 using GeneratorDuty.Repository;
-using GeneratorDuty.Services;
 using Telegram.Bot;
 using Telegram.Bot.Types;
-namespace GeneratorDuty.CallBackKeyboards;
 
-public class DutyReject(DutyRepository repository, MemoryExceptionDuty cache) : CallQuery
+namespace GeneratorDuty.CallBackKeyboards.Duty;
+
+public class DutyAccept(DutyRepository repository) : CallQuery
 {
-    public override string Name { get; set; } = "duty_reject";
+    public override string Name { get; set; } = "duty_accept";
 
     public override async void Execute(ITelegramBotClient client, CallbackQuery callbackQuery)
     {
         var array = TryGetArrayFromCallBack(callbackQuery);
         if (callbackQuery.Message is null || array is null || array.Length == 0 ||
             !long.TryParse(array[0], out var idMemberDuty)) return;
-
+        
         if (Restrictions.ChatIdsRequiredAdminRights.Contains(callbackQuery.Message.Chat.Id) && !await client.IsUserAdminInChat(callbackQuery.From.Id, callbackQuery.Message.Chat.Id))
         {
             await client.AnswerCallbackQueryAsync(callbackQuery.Id, "❌ \n\nВ этом чате данное действие могут выполнять только админы беседы", true);
@@ -28,19 +27,16 @@ public class DutyReject(DutyRepository repository, MemoryExceptionDuty cache) : 
         var memberDuty = await repository.Members.GetMemberDuty(idMemberDuty);
         if (memberDuty is null) return;
 
-        cache.AddMemberDuty(memberDuty);
-
-        await repository.LogsMemberPriority.Create(new LogDutyMemberPriority
+        await repository.LogsMembers.Create(new LogDutyMember
         {
             UserId = memberDuty.Id,
+            Date = DateTime.Now
         });
-
-        await client.TrySendMessage(callbackQuery.Message.Chat.Id,
-            $"О как.. Я запомнил, что {memberDuty.MemberNameDuty} сегодня нет. В следующий раз, заставлю отдежурить 😈");
         
-        await client.TryDeleteMessage(callbackQuery.Message.Chat.Id, callbackQuery.Message.MessageId);
-
-        GetCommand command = new(repository, cache);
-        await command.ExecuteAsync(client,callbackQuery.Message);
+        foreach (var member in await repository.LogsMemberPriority.GetLogsByIdMember(idMemberDuty))
+            await repository.LogsMemberPriority.Remove(member);
+        
+        await client.EditMessageReplyMarkupAsync(callbackQuery.Message.Chat.Id, callbackQuery.Message.MessageId,
+            replyMarkup: null);
     }
 }
