+++ GeneratorDuty/CallBackKeyboards/Duty/DutyReject.cs
@@ -1,42 +1,47 @@
-﻿using GeneratorDuty.Common;
+﻿using GeneratorDuty.Commands;
+using GeneratorDuty.Common;
 using GeneratorDuty.CustomRights;
 using GeneratorDuty.Extensions;
 using GeneratorDuty.Models;
 using GeneratorDuty.Repository;
+using GeneratorDuty.Services;
 using Telegram.Bot;
 using Telegram.Bot.Types;
 
-namespace GeneratorDuty.CallBackKeyboards;
+namespace GeneratorDuty.CallBackKeyboards.Duty;
 
-public class DutyForce(DutyRepository repository) : CallQuery
+public class DutyReject(DutyRepository repository, MemoryExceptionDuty cache) : CallQuery
 {
-    public override string Name { get; set; } = "duty_force";
+    public override string Name { get; set; } = "duty_reject";
 
     public override async void Execute(ITelegramBotClient client, CallbackQuery callbackQuery)
     {
         var array = TryGetArrayFromCallBack(callbackQuery);
         if (callbackQuery.Message is null || array is null || array.Length == 0 ||
             !long.TryParse(array[0], out var idMemberDuty)) return;
-        
+
         if (Restrictions.ChatIdsRequiredAdminRights.Contains(callbackQuery.Message.Chat.Id) && !await client.IsUserAdminInChat(callbackQuery.From.Id, callbackQuery.Message.Chat.Id))
         {
             await client.AnswerCallbackQueryAsync(callbackQuery.Id, "❌ \n\nВ этом чате данное действие могут выполнять только админы беседы", true);
             return;
         }
         
-        var members = await repository.Members.GetMemberDuty(idMemberDuty);
-        if (members is null) return;
+        var memberDuty = await repository.Members.GetMemberDuty(idMemberDuty);
+        if (memberDuty is null) return;
 
-        await repository.LogsMembers.Create(new LogDutyMember
+        cache.AddMemberDuty(memberDuty);
+
+        await repository.LogsMemberPriority.Create(new LogDutyMemberPriority
         {
-            UserId = members.Id,
-            Date = DateTime.Now
+            UserId = memberDuty.Id,
         });
-        
-        foreach (var member in await repository.LogsMemberPriority.GetLogsByIdMember(idMemberDuty))
-            await repository.LogsMemberPriority.Remove(member);
+
+        await client.TrySendMessage(callbackQuery.Message.Chat.Id,
+            $"О как.. Я запомнил, что {memberDuty.MemberNameDuty} сегодня нет. В следующий раз, заставлю отдежурить 😈");
         
         await client.TryDeleteMessage(callbackQuery.Message.Chat.Id, callbackQuery.Message.MessageId);
-        await client.TrySendMessage(callbackQuery.Message.Chat.Id,$"Назначен дежурный вручную: {members.MemberNameDuty}");
+
+        GetCommand command = new(repository, cache);
+        await command.ExecuteAsync(client,callbackQuery.Message);
     }
 }
